A4Q3 Finished

+++ src/Assignment4/A4Q3.java
@@ -21,10 +21,10 @@ public class A4Q3 {
         Scanner input = new Scanner(System.in);
         System.out.println("Please enter four numbers on seperate lines");
         // telling the system that there are 4 numbers are needed in the system
-        double Number1 = input.nextInt();
-        double Number2 = input.nextInt();
-        double Number3 = input.nextInt();
-        double Number4 = input.nextInt();
+        double Number1 = input.nextDouble();
+        double Number2 = input.nextDouble();
+        double Number3 = input.nextDouble();
+        double Number4 = input.nextDouble();
 
         // once the numbers are put in they are all put into one line with commas seperating them
         System.out.println("Your numbers were" + Number1 + "," + Number2 + "," + Number3 + "and" + Number4);
